requirements.txt
# Requirements automatically generated by pigar.
# https://github.com/damnever/pigar

# analyze.py: 3
# fetch_lyrics.py: 1
# main.py: 3
PyYAML == 3.13

# analyze.py: 2
# main.py: 2
nltk == 3.4

# fetch_lyrics.py: 2
requests == 2.18.4
